tests/wpt/web-platform-tests/app-history/transition/location-href-transitionWhile-rejected.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
async_test(t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  window.onload = () => t.step_timeout(() => {
    assert_equals(appHistory.transition, null);
    const from = appHistory.current;

    appHistory.addEventListener("navigate", t.step_func(e => {
      e.transitionWhile(Promise.reject(new Error("test")));
      assert_equals(appHistory.transition, null);
    }));

    appHistory.addEventListener("navigateerror", t.step_func(() => {
      assert_equals(appHistory.transition.navigationType, "push");
      assert_equals(appHistory.transition.from, from);

      Promise.resolve().then(t.step_func_done(() => {
        assert_equals(appHistory.transition, null);
      }));
    }));

    location.href = "?1";

    assert_equals(appHistory.transition.navigationType, "push");
    assert_equals(appHistory.transition.from, from);
  }, 0);
}, "transitionWhile()ed location.href navigation with a rejected promise");
</script>
